deleted template language div box
<div class="alert alert-danger">
This is the workshop template. Delete these lines and use it to
<a href="https://carpentries.github.io/workshop-template/customization/index.html">customize</a>
your own website. If you are running a self-organized workshop or have not put
in a workshop request yet, please also fill in
<a href="{{site.amy_site}}/forms/self-organised/">this workshop request form</a>
to let us know about your workshop and our administrator may contact you if we
need any extra information.
If this is a pilot workshop for a new lesson,
set the `pilot` field to `true` in `_config.yml`.
For workshops teaching a lesson in The Carpentries Incubator,
remember to uncomment the `incubator_lesson_site`, `incubator_pre_survey`, and `incubator_post_survey`
fields in `_config.yml`
</div>